+++ CHANGES.md
@@ -3,9 +3,20 @@
 ### ???? - ????
 
 ##### Additions :tada:
+
 - `UCesiumGltfPrimitiveComponent` now has static mobility, allowing it to finally take advantage of several rendering optimizations only available for static objects.
 - Added `UCesiumBoundingVolumePoolComponent` and `UCesiumBoundingVolumeComponent` to track the occlusion state of prospective tiles before they are loaded. This allows tile loads to be avoided if they will be occluded anyways.
 - Added options in `ACesium3DTileset` to control occlusion culling and turn it off if necessary. 
+- Added a new delegate `OnTilesetLoaded` that is exposed to the blueprint, which is invoked when the current tileset has finished loading.
+- Added a getter `GetLoadProgress` exposed to blueprint, which returns the current loaded percentage of the tileset. 
+- Added BP accessible callback `OnFlightComplete` for when Dynamic Pawn completes flight. 
+- Added BP accessible callback `OnFlightInterrupt` for when Dynamic Pawn's flying is interrupted. 
+
+##### Fixes :wrench:
+
+- Perforce marks files read only if not checked out. To save access tokens, it now asks to check out the necessary config files from source control.
+- Fixed credit images not appearing in UE5.
+- Added `UPROPERTY` to the Credit Widget to prevent it from being deleted by the garbage collector. 
 
 ### v1.15.0 - 2022-07-01
 
@@ -21,6 +32,9 @@
 - Fixed a bug that could cause a crash when reporting tileset or raster overlay load errors, particularly while switching levels.
 - We now Log the correct asset source when loading a new tileset from either URL or Ion. 
 - Disabling physics meshes of a tileset now works in Unreal Engine 5. 
+
+##### Breaking Changes :mega:
+- Add option to scale the viewport by pixel density to increase performance for mobile devices.
 
 ### v1.14.0 - 2022-06-01